Remove the explicit specification of aws_appautoscaling_target.role_arn
At the moment, when deploying the Terraform for the PSS module,
the aws_appautoscaling_target.service_autoscaling_target always
wants to be updated because the role we specify is different to
the one which AWS is actually using.

This field is optional, and will be populated automatically by AWS
depending on the service being autoscaled.

Because AWS is populating the field for us, this is causing drift
between the state file which is what we think the value is/should be,
and what the actual value is.

Letting AWS manage this for us prevents this repeated update, and
reduces the complexity of our Terraform code.